+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.13.0] - 2026-06-15
+
+### Added
+- Added support for configuring Tagref with a `tagref.yml` file.
+- Added the `--config` option for specifying the location of the config file.
+- Added support for configuring extra ignore rules in `tagref.yml`.
+
+### Removed
+- Removed the command-line options for scan paths and sigils. Sigils are now configured in
+  `tagref.yml`.
+
 ## [1.12.1] - 2026-04-13
 
 ### Fixed